+++ backslash/test.py
@@ -7,13 +7,14 @@ from .related_entity_container import RelatedEntityContainer
 from .warning_container import WarningContainer
 from .lazy_query import LazyQuery
 from .metadata_holder import MetadataHolder
+from .timing_container import TimingContainer
 
 
-class Test(APIObject, MetadataHolder, ErrorContainer, WarningContainer, Commentable, RelatedEntityContainer):
+class Test(APIObject, MetadataHolder, ErrorContainer, WarningContainer, Commentable, RelatedEntityContainer, TimingContainer):
 
     @property
     def ui_url(self):
-        return self.client.url + '#/sessions/{}/tests/{}'.format(self.session_display_id, self.logical_id or self.id)
+        return self.client.get_ui_url('sessions/{}/tests/{}'.format(self.session_display_id, self.logical_id or self.id))
 
     def report_end(self, duration=NOTHING):
         self.client.api.call_function('report_test_end', {'id': self.id, 'duration': duration})
@@ -23,9 +24,6 @@ class Test(APIObject, MetadataHolder, ErrorContainer, WarningContainer, Commenta
 
     def report_interrupted(self):
         self.client.api.call_function('report_test_interrupted', {'id': self.id})
-
-    def edit_status(self, status):
-        return self.client.api.call_function('edit_test_status', {'id': self.id, 'status': status})
 
     def query_errors(self):
         """Queries tests of the current session
@@ -39,3 +37,6 @@ class Test(APIObject, MetadataHolder, ErrorContainer, WarningContainer, Commenta
 
     def get_parent(self):
         return self.get_session()
+
+    def update_status_description(self, description):
+        return self.client.api.call_function('update_status_description', {'test_id': self.id, 'description': description})