+++ docs/changelog.rst
@@ -1,12 +1,17 @@
 Changelog
 =========
 
+* :release:`2.32.0 <30-10-2017>`
+* :feature:`60` Clean up UI URL generation, added ``Backslash.get_ui_url`` helper method
+* :feature:`59` Support reporting interruption exceptions to Backslash
+* :feature:`58` Support reporting timing metrics
+* :feature:`57` Support reporting test status description
 * :release:`2.31.2 <14-9-2017>`
 * :bug:`54` Handle cases of detached head correctly when deducing local branch
 * :release:`2.31.1 <11-9-2017>`
 * :bug:`53` Use api session when constructing lazy queries
 * :release:`2.31.0 <10-9-2017>`
-* :feature:`52` Support reporting sessions with a specific TTL, marking them for future deletion on the server
+* :feature:`52` Support reporting sessions with a specific TTL, marking them for future deletion on the server. This can be also specified in the command-line, by passing ``--session-ttl-days=X``
 * :feature:`51` Report local and remote SCM branches if supported
 * :release:`2.30.0 <8-8-2017>`
 * :feature:`50` Added session_webapp_url property to the Slash plugin